requirements.txt
asgiref==3.5.2
Django==3.0
django-autoslug==1.9.8
django-crispy-forms==1.14.0
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
